+++ code/game/ai_main.c
@@ -607,7 +607,7 @@ void BotInputToUserCommand(bot_input_t *bi, usercmd_t *ucmd, int delta_angles[3]
 	ucmd->forwardmove = DotProduct(forward, bi->dir) * bi->speed;
 	ucmd->rightmove = DotProduct(right, bi->dir) * bi->speed;
 	// Was always 0 in basejk
-	ucmd->upmove = fabs(forward[2]) * bi->dir[2] * bi->speed;
+	ucmd->upmove = fabsf(forward[2]) * bi->dir[2] * bi->speed;
 	//normal keyboard movement
 	if (bi->actionflags & ACTION_MOVEFORWARD) ucmd->forwardmove += 127;
 	if (bi->actionflags & ACTION_MOVEBACK) ucmd->forwardmove -= 127;
